Explicitly add a survey method to UserController
Previously the survey.rhtml method was rendered by default
if no action matching 'survey' was found in the controller,
but this doesn't seem to be the case with some configurations.
This is rather unsatisfactory, in that we haven't found
documentation of any change that might cause this different
behaviour, but the pragmatic solution is to just to add
the empty survey method and move on.