Skip lib/censor_rules if the relevant tables don't exist
With the split deploy scripts, we now do database migration after the
theme installation - this is fine apart from on the first deployment
where the info_requests and censor_rules don't exist yet.  So, skip
the creation of the censor rules if those tables aren't present.

This will mean that a new instance of WDTK when first deployed won't
have these default censor rules, but de novo instances of WDTK (other
than dev sites) are rare nowadays, so this probably isn't a big issue.