Update public_body_categories_en.rb
An extension to FOISA came into force recently on Tuesday 1 April.

My understanding of it is that it makes bodies established or created solely by one or more local authorities whose functions on behalf of any of those authorities (include developing and/or delivering recreational, sporting, cultural or social facilities and activities) and which in carrying out those functions is financed wholly or in part by any of those authorities subject to the Freedom of Information (Scotland) Act 2002.