Include cautionary new request text for pro users
Copies the standard warnings about requests being a matter of public
record and not requesting personal information when making new requests
to the pro version of the new request sidebar.

The sentence about requests being public has been modified slightly to
reflect that requests may initially be made in private.