Extract follow button and count into mixin
Fixes the alignment issue described in #418 and abtracts the code into a
mixin to cut down on reuse.

Before when this code was duplicated across the stylesheet in one of the
copies there was a 'font-size: 1em;' declaration [1].

This has been removed as part of turning it into a mixin.

[1] https://github.com/mysociety/whatdotheyknow-theme/blob/b0876fe/app/assets/stylesheets/responsive/custom.scss#L976