Update links to the Charity Commission and Charity Register sites
Both sites have updated their URL structure so our current links to
registration pages of charities no longer work. This commit adopts the
current URL structures, although this may be fragile in the case of the
Charity Commission as they are currently linking to a beta subdomain
with a banner indicating that it is a prototype.