A/B test position of donate-cta on request page
Moves the red donation banner from the sidebar to the end of the page,
using JavaScript, if the current user is in Context Experiment group 1.

Also tracks clicks on the donation banner button, no matter where it’s
located, so we can use that as the goal in our Content Experiment.

See mysociety/whatdotheyknow-theme#542 and (more specifically)
mysociety/alaveteli-experiments#115.