Add section regarding requests for information about deceased persons
Adds information about the scenarios where it may be appropriate to use WhatDoTheyKnow to request information about a deceased person; and where it might be more appropriate to contact a public body privately.

This also provides a link to the new guidance in 9ffcccc  regarding removal of information on compassionate grounds.