Add exemption titles to wizard labels
Add short title to each exemption for better user understanding.

I used a hyphen rather than a colon because after processing the ERB the
YAML interpreter was expecting the whole string to be quoted. This means
we'd have to do something like `<%= _(%q("Section 12: Foo")) %>` which
is pretty gross.

Ideally these labels would support HTML rendering so that we could bold
the section number, but this is the easiest fix for now.

Fixes https://github.com/mysociety/alaveteli/issues/6237