Hack redirecting to ICO advice
We were originally intending to redirect to a separate help page for
advice on referring a request to the ICO, but we decided that keeping
the advice on a single page is simpler for users.

We don't have an `anchor:` target, but we can use the `external`
mechanism to hack it in.

This results in a new page load, but it does get the user to the right
place. We can later improve by adding an anchor target.