Prevent disabling of Internal Admin email alerts
For some reason we keep seeing email alerts getting disabled for the
Internal Admin User on WhatDoTheyKnow.

We think this is somehow related to the Unsubscribe link rendered via
`app/views/track_mailer/event_digest`. Perhaps someone is mistakenly
clicking it, or an email client is pre-emptively clicking it to cache
content? Who knows!?

This commit checks that the user is _not_ the Internal Admin before
disabling. I used a new module with `prepend` so that we can call
`super` to rely on core behaviour, rather than having to keep the main
method definition in sync.

Fixes https://github.com/mysociety/whatdotheyknow-theme/issues/931.