Set theme-color meta tag
Makes the browser URL bar a nice colour [1]. Uses a pro/non-pro
appropriate colour by checking the user. Ideally we'd interpolate the
SCSS setting but don't think we have an easy way of doing that right
now.

In the case of the Pro theme, the darker colour looked better than the
lighter part of the gradient.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meta/name/theme-color