Update about.html.erb
(1) At a time when the cost of living is increasing in the UK and major national newspapers are talking about a cost of living crisis, it might be helpful to make it even clearer that people can use the site for free.

(2) We are all about transparency so making it really clear at the outset that you can make FOI requests for free is in line with our principles.

(3) In behavioural economics, there is something called 'the Power of Free'. My limited understanding of this concept is that people have a very high preference for things that are free which is higher than you might otherwise expect even by comparisons to things that cost almost nothing e.g. one penny (GBP). If I have got this right then even of the people who think the cost of using the site is likely be very low and easily affordable the knowledge that it is free will encourage a higher proportion to sign up.

https://www.forbes.com/sites/piyankajain/2018/03/01/5-behavioral-economics-principles-for-marketeers/?sh=51c10c8e28eb