Add basic error handling for research feed
We don't want the homepage to error just because there's an error or
timeout trying to pull in research articles, so add some error handling
to only render if we have articles to render.

Ideally we'd do some clever caching like we do for the AskTheEU homepage
blog cache [1] to pull from cached versions in the event of an error,
but that's not yet available in Alaveteli core and isn't generalised
enough for this.

[1] https://github.com/mysociety/asktheeu-theme/pull/44