Add caching to frontpage research papers
While the frontpage is cached for 5 minutes, we don't need to fetch the
research papers quite so frequently. This commit caches the papers for
an hour, so that we only get one potentially slow request per hour.

This might result in not displaying any papers for an hour if the feed
fails, but that's fine as this isn't critical content.

Ideally we wouldn't fetch this in a request cycle at all.