Add all the glossary letters to the nav
It's a bit disorienting when you're looking for a letter and it's not
next to the letter you were expecting because of the jump (e.g. I'm
looking for "L", expecting it to be right of "K", but it's actually
right of "I").

This commit adds missing letters, styling them with a muted colour so
it's obvious there's no entries for that letter.