Add help/unhappy to sidebar
Though this page is part of an in-app workflow, it contains good general
advice that stands on its own. This links it from the sidebar so that
people can get to it.

As part of this I've moved the no-response page a bit lower down. I
think it makes sense to group like:

* About the ATI acts
* Exemptions (subtleties related to the above)
* Poor experience with using the acts
* Glossary – reference material