Add Pro sign up ban check
When a YAML config file is present, check the payment card before the
pro subscription is created.

With the YAML:

```yaml
- fingerprint: 123abc456DEF789gh
  address_zip: AB11AB
- address_zip: CD11CD
```

This would add two bans. Prevent sign up if the payment card entered
matches:
1. the fingerprint and postcode `AB1 1AB`,
or
2. the postcode `cd11cd`

For fingerprint, these come from the Stripe dashboard and are unique for
a given card. See: https://stripe.com/docs/api/cards/object#card_object-fingerprint

For postcodes/address zip we upcase and strip all white space from what
is provided on the payment form.