WIP: Add citation promo to pro mailouts
Adding only the HTML part to the theme seems to prevent the mail being
sent as multipart. IDK whether this is an issue with the rails previewer
not working well with our theme override, or whether this will also
happen in production with it stopping at the theme `embargo_mailer`
directory instead of falling back to the core one to get the text
version. If so, we'll just have to override the text template without
modification.

Obviously need to add the link to the blog post, and maybe tweak
phrasing / design of the banner.