[WDTK] Improved visual elements for button classes
Currently the hover effect is quite subtle, which is not ideal for users with visual impairments. Ideally we want to change at least two properties when changing between states, to make it fairly obvious. For the tertiary button, I also changed the default state, to make more evident that is an interactive element. For this I increased the border-width, and changed the color(border and font) to what we use for links, which makes it more familiar for users that is an interactive element.