Added skip to main content link
At the moment if an user wants to use the keyboard to navigate, they have to tab through all the links in the sidebar, which can take a lot of time. Also considering the sidebar always has the same/similar links seems like a pain if the are navigation other help pages.