WIP Fix PII badger background job
Sometimes this is breaking for files which have redactions/masks.

In the case of some XLSX files text masking of email addresses in the
underlying ZIP format meant the Python Excel Analyzer code couldn't
decompress the file and analyse the file contents.

The patch for the ActiveStorage mirror service is needed for our server
deployment as these jobs run on a secondary application instance which
doesn't have the raw email on disk, so we have to rely on fetching the
S3 mirrored version. This patch should probably be moved elsewhere.