Improve styling of FOI Myths on Learn page
Headings elsewhere in the Learn materials state facts. Which made this
myths page confusing, as you instinctively want to trust what’s marked
up as a heading, but here, the headings were all lies!

This new styling wraps the myths in a red box (to imply “wrongness”)
and instead uses the “correction” immediately after each myth as the
heading, to avoid misinterpretation when skim-reading the page.